+++ apps/react-demo/vite.config.ts
@@ -8,7 +8,7 @@ export default defineConfig(() => ({
   root: __dirname,
   cacheDir: '../../node_modules/.vite/apps/react-demo',
   server: {
-    port: 4200,
+    port: 4201,
     host: 'localhost',
   },
   preview: {