+++ packages/angular/.storybook/main.ts
@@ -1,0 +1,25 @@
+import type { StorybookConfig } from '@storybook/angular';
+
+const config: StorybookConfig = {
+  stories: ['../**/*.@(mdx|stories.@(js|jsx|ts|tsx))'],
+  addons: ['@storybook/addon-docs'],
+  framework: {
+    name: '@storybook/angular',
+    options: {
+      builder: {
+        viteConfigPath: 'vite.config.mts',
+      },
+    },
+  },
+  staticDirs: [
+    // Map the local themes directory to a public path
+    { from: '../../themes/src/themes', to: '/themes' }
+  ],
+}
+
+export default config
+
+// To customize your Vite configuration you can use the viteFinal field.
+// Check https://storybook.js.org/docs/react/builders/vite#configuration
+// and https://nx.dev/recipes/storybook/custom-builder-configs
+