+++ packages/angular/.storybook/themes.ts
@@ -1,0 +1,8 @@
+/**
+ * Themes that should be available in storybook
+ */
+
+export const themes = [
+    { name: "Forvaltning", href: "themes/forvaltning.css" },
+    { name: "Ledsagerbevis", href: "themes/ledsagerbevis.css" },
+];