+++ packages/angular/src/components/checkbox/checkbox.ts
@@ -1,4 +1,4 @@
-import { Component, model } from '@angular/core'
+import { booleanAttribute, Component, input, model } from '@angular/core'
 import { CommonInputs } from '../common-inputs'
 
 @Component({
@@ -12,10 +12,19 @@ import { CommonInputs } from '../common-inputs'
     class: 'ds-input',
     '[attr.id]': 'id()',
     '[attr.aria-describedby]': 'ariaDescribedBy()',
+    '[attr.readonly]': 'readonly() ? true : null',
+    '(click)': 'onClick($event)',
   },
   template: ``,
 })
 export class Checkbox {
   id = model<string>()
-  ariaDescribedBy = model<string>()
+  ariaDescribedBy = model<string | undefined>(undefined, { alias: 'aria-describedby' })
+  readonly = input(false, { transform: booleanAttribute })
+
+  onClick(event: Event) {
+    if (this.readonly()) {
+      event.preventDefault()
+    }
+  }
 }