+++ packages/angular/src/utils/log-if-devmode.ts
@@ -1,0 +1,7 @@
+import { isDevMode } from "@angular/core";
+
+export const logIfDevMode = ({ component, message }: { component: string; message: string; }) => {
+    if (isDevMode()) {
+        console.log(`[${component}] ${message}`)
+    }
+}